Создает абстрактный класс. Добавляет наследование простым компонентам
