+++ src/view/list-points-view.js
@@ -1,26 +1,12 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createListPointsTemplate = () => (
   `<ul class="trip-events__list">
   </ul>`
 );
 
-export default class ListPointsView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class ListPointsView extends AbstractView {
   get template() {
     return createListPointsTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
