+++ src/view/site-menu-view.js
@@ -1,5 +1,5 @@
 import {MENU_ITEMS} from '../const.js';
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const activeItem = MENU_ITEMS[0];
 
@@ -20,22 +20,8 @@ const createSiteMenuTemplate = () => {
 </nav>`;
 };
 
-export default class SiteMenuView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class SiteMenuView extends AbstractView {
   get template() {
     return createSiteMenuTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
